compile when qreal is float and not double
svn path=/branches/KDE/4.1/kdegraphics/gwenview/; revision=876218

+++ lib/imageview.cpp
@@ -621,7 +621,7 @@ qreal ImageView::computeZoomToFit() const {
 	}
 
 	if (!d->mEnlargeSmallerImages) {
-		zoom = qMin(zoom, 1.0);
+		zoom = qMin(zoom, qreal(1.0));
 	}
 
 	return zoom;
